Changed view interface to the same file of ranking activity * Renamed RankingActivity.kt to RankingView.kt
